:construction: Separate directories for each function
