+++ animus/cmd/main.go
@@ -7,7 +7,7 @@ import (
 	"github.com/GoogleCloudPlatform/functions-framework-go/funcframework"
 	_ "github.com/GoogleCloudPlatform/functions-framework-go/funcframework"
 	// Blank-import the function package so the init() runs
-	_ "github.com/KasumiMercury/patotta-stone-functions-go"
+	_ "github.com/KasumiMercury/patotta-stone-functions-go/animus"
 )
 
 func main() {
