+++ animus/pkg/model/chat.go
