+++ animus/pkg/usecase/chat.go
@@ -4,9 +4,9 @@ import (
 	"context"
 	"encoding/json"
 	"fmt"
-	"github.com/KasumiMercury/patotta-stone-functions-go/pkg/model"
-	"github.com/KasumiMercury/patotta-stone-functions-go/pkg/repository"
-	"github.com/KasumiMercury/patotta-stone-functions-go/pkg/service"
+	model2 "github.com/KasumiMercury/patotta-stone-functions-go/animus/pkg/model"
+	"github.com/KasumiMercury/patotta-stone-functions-go/animus/pkg/repository"
+	"github.com/KasumiMercury/patotta-stone-functions-go/animus/pkg/service"
 	"log/slog"
 	"os"
 	"slices"
@@ -33,7 +33,7 @@ func NewChatUsecase(targetChannel []string, chatSvc service.Chat, supaRepo repos
 func (u *chatUsecase) FetchChatsFromStaticTargetVideo(ctx context.Context) error {
 	// load info of static target video from environment variables
 	stcEnv := os.Getenv("STATIC_TARGET")
-	var stc model.VideoInfo
+	var stc model2.VideoInfo
 	if err := json.Unmarshal([]byte(stcEnv), &stc); err != nil {
 		slog.Error("Failed to unmarshal STATIC_TARGET", "error", err)
 		// If the environment variable is not set correctly, the function will panic.
@@ -70,8 +70,8 @@ func (u *chatUsecase) FetchChatsFromStaticTargetVideo(ctx context.Context) error
 	return nil
 }
 
-func filterChatsByAuthorChannel(chats []model.YTChat, targetChannel []string) ([]model.YTChat, []model.YTChat) {
-	var targetChats, otherChats []model.YTChat
+func filterChatsByAuthorChannel(chats []model2.YTChat, targetChannel []string) ([]model2.YTChat, []model2.YTChat) {
+	var targetChats, otherChats []model2.YTChat
 	for _, chat := range chats {
 		if slices.Contains(targetChannel, chat.AuthorChannelID) {
 			targetChats = append(targetChats, chat)
@@ -83,7 +83,7 @@ func filterChatsByAuthorChannel(chats []model.YTChat, targetChannel []string) ([
 	return targetChats, otherChats
 }
 
-func (u *chatUsecase) filterChatsByPublishedAt(ctx context.Context, chats []model.YTChat, sourceId string) ([]model.YTChat, error) {
+func (u *chatUsecase) filterChatsByPublishedAt(ctx context.Context, chats []model2.YTChat, sourceId string) ([]model2.YTChat, error) {
 	// Fetch the last recorded chat's publishedAt from the Supabase
 	threshold, err := u.supaRepo.GetPublishedAtOfLastRecordedChatBySource(ctx, sourceId)
 	if err != nil {
@@ -94,7 +94,7 @@ func (u *chatUsecase) filterChatsByPublishedAt(ctx context.Context, chats []mode
 	// Filter the chats by the threshold
 	// The chats are already sorted by the publishedAt in ascending order (constraint of the YouTube API)
 
-	var filteredChats []model.YTChat
+	var filteredChats []model2.YTChat
 
 	for i, chat := range chats {
 		if chat.PublishedAtUnix > threshold {
